Rename package: observation > observationunits
Issue: IBP-3213
Reviewer: None
